move(cosmo): move io to private API
fixing io docstrings that were untested
