+++ astropy/cosmology/_src/io/builtin/latex.py
