+++ astropy/cosmology/_src/io/builtin/model.py
