+++ astropy/cosmology/_src/io/builtin/utils.py
