+++ astropy/cosmology/_src/tests/io/test_ecsv.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from astropy.cosmology._io.ecsv import read_ecsv, write_ecsv
+from astropy.cosmology._src.io.builtin.ecsv import read_ecsv, write_ecsv
 from astropy.cosmology.core import _COSMOLOGY_CLASSES
 from astropy.table import QTable, Table, vstack
