+++ astropy/cosmology/_src/tests/io/test_html.py
@@ -3,7 +3,11 @@
 import pytest
 
 import astropy.units as u
-from astropy.cosmology._io.html import _FORMAT_TABLE, read_html_table, write_html_table
+from astropy.cosmology._src.io.builtin.html import (
+    _FORMAT_TABLE,
+    read_html_table,
+    write_html_table,
+)
 from astropy.table import QTable, Table, vstack
 from astropy.utils.compat.optional_deps import HAS_BS4
