+++ astropy/cosmology/_src/tests/io/test_table.py
@@ -3,7 +3,7 @@
 import pytest
 
 from astropy.cosmology import Cosmology
-from astropy.cosmology._io.table import from_table, to_table
+from astropy.cosmology._src.io.builtin.table import from_table, to_table
 from astropy.cosmology.core import _COSMOLOGY_CLASSES
 from astropy.table import QTable, Table, vstack
