+++ astropy/cosmology/_src/tests/io/test_yaml.py
@@ -5,7 +5,7 @@ import pytest
 import astropy.units as u
 from astropy.cosmology import Cosmology, FlatLambdaCDM, Planck18
 from astropy.cosmology import units as cu
-from astropy.cosmology._io.yaml import (
+from astropy.cosmology._src.io.builtin.yaml import (
     from_yaml,
     to_yaml,
     yaml_constructor,
@@ -20,7 +20,7 @@ from .base import ToFromDirectTestBase, ToFromTestMixinBase
 
 
 def test_yaml_representer():
-    """Test :func:`~astropy.cosmology._io.yaml.yaml_representer`."""
+    """Test :func:`~astropy.cosmology._src.io.builtin.yaml.yaml_representer`."""
     # test function `representer`
     representer = yaml_representer("!astropy.cosmology.flrw.LambdaCDM")
     assert callable(representer)
@@ -32,7 +32,7 @@ def test_yaml_representer():
 
 
 def test_yaml_constructor():
-    """Test :func:`~astropy.cosmology._io.yaml.yaml_constructor`."""
+    """Test :func:`~astropy.cosmology._src.io.builtin.yaml.yaml_constructor`."""
     # test function `constructor`
     constructor = yaml_constructor(FlatLambdaCDM)
     assert callable(constructor)
